fix(specs) big refactor

+++ app/controllers/answers_controller.rb
@@ -1,18 +1,16 @@
 class AnswersController < ApplicationController
 
-  def index
-    redirect_to question
-  end
+  before_action :authenticate_user!
 
   def edit; end
 
   def create
     @answer = question.answers.new(answer_params)
     @answer.author = current_user
-    if @answer.save
-      redirect_to answer.question
+    if @answer.save && current_user
+      redirect_to @answer.question
     else
-      render '/questions/show'
+      render 'questions/show'
     end
   end
 
@@ -25,8 +23,8 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    answer.destroy
-    redirect_to question_answers_path(answer.question)
+    answer.destroy if current_user.author_of?(answer)
+    redirect_to question_path(answer.question)
   end
 
   private
